+++ online-store/src/components/index.ts
@@ -1,0 +1,4 @@
+export * from './Basket';
+export * from './filter';
+export * from './search';
+export * from './sort';
